Use environment variables for controller configuration (#1139)
**Yet another PR split out from #1133 to try reduce review burden** -
keeping that one open for now as all of these other PRs are ultimately
working towards that goal.

This PR refactors the kagent controller to support the use of
environment variables for configuration in addition to command-line
arguments. It also updates the Helm chart to make use of env vars
instead of command line args and adds the ability for user's to supply
their own environment variables with custom configuration. This allows
users to supply sensitive configuration (e.g. postgres database url) via
secrets instead of exposing these via `args`. Env vars are also easier
to patch when working with rendered manifests if needed.

---------

Signed-off-by: Brian Fox <878612+onematchfox@users.noreply.github.com>
Signed-off-by: Ivan Porta <porta.ivan@outlook.com>